2 separate bidi routes
